pyproject.toml
[project]
name = "chatbot-memory"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "langchain>=0.3.27",
    "langchain-core>=0.3.75",
    "langchain-groq>=0.3.7",
    "langgraph>=0.6.7",
    "python-dotenv>=1.1.1",
    "streamlit>=1.49.1",
]
